refactor: move expectation-policy related concepts into expectation/Common.hpp
