+++ test/unit-tests/Sequence.cpp
@@ -1,10 +1,9 @@
-//          Copyright Dominic (DNKpp) Koepke 2024 - 2025.
+//          Copyright Dominic (DNKpp) Koepke 2024-2026.
 // Distributed under the Boost Software License, Version 1.0.
 //    (See accompanying file LICENSE_1_0.txt or copy at
 //          https://www.boost.org/LICENSE_1_0.txt)
 
 #include "mimic++/Sequence.hpp"
-#include "mimic++/Expectation.hpp"
 #include "mimic++/utilities/C++23Backports.hpp"
 
 #include "TestReporter.hpp"
