+++ test/unit-tests/expectation/CMakeLists.txt
@@ -1,0 +1,8 @@
+#          Copyright Dominic (DNKpp) Koepke 2026.
+# Distributed under the Boost Software License, Version 1.0.
+#    (See accompanying file LICENSE_1_0.txt or copy at
+#          https://www.boost.org/LICENSE_1_0.txt)
+
+target_sources(${TARGET_NAME} PRIVATE
+    "Common.cpp"
+)
